Move server side exception to the separate package
