+++ vaadin-connect/src/main/java/com/vaadin/connect/VaadinConnectController.java
@@ -35,7 +35,6 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import com.vaadin.connect.auth.VaadinConnectAccessChecker;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -50,6 +49,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.vaadin.connect.auth.VaadinConnectAccessChecker;
+import com.vaadin.connect.exception.VaadinConnectException;
 
 /**
  * The controller that is responsible for processing Vaadin Connect requests.
